+++ addon/-private/utils/is-fastboot.js
